move generator functions into struct and module
