+++ native/src/open_api/mod.rs
@@ -1,17 +1,14 @@
-mod add_paths;
 mod generator;
 mod generator_v2;
 mod open_api;
 mod symbol_table_helpers;
 
-use std::{collections::BTreeMap, fs::File, io::Write, path::PathBuf};
+use std::{fs::File, io::Write, path::PathBuf};
 
 use neon::{prelude::*, result::Throw};
 use serde_json::json;
 
-use crate::typescript::DeclarationTable;
-
-use self::{generator_v2::from_source_file, open_api::OpenApi};
+use self::{generator_v2::GeneratorV2, open_api::OpenApi};
 
 fn merge_schemas(
     open_api: &OpenApi,
@@ -43,13 +40,13 @@ fn merge(target: &mut serde_json::Value, overlay: &serde_json::Value) {
 }
 
 fn generate_schema(open_api_handle: Handle<JsObject>, cx: &mut FunctionContext) -> Result<String, Throw> {
+    let mut generator = GeneratorV2::new();
     let paths = open_api_handle.get::<JsArray, FunctionContext, &str>(cx, "paths")?;
-    let mut symbol_tables = BTreeMap::<String, DeclarationTable>::new();
 
     let mut result = OpenApi::new();
     for path in paths.to_vec(cx)? {
         let path = path.downcast_or_throw::<JsString, _>(cx)?.value(cx);
-        let open_api = from_source_file(path, &mut symbol_tables);
+        let open_api = generator.from_source_file(&path);
         result.merge(open_api);
     }
